refactor: post merge cleanup
