+++ src/util/files.ts
@@ -3,8 +3,6 @@ import path from 'path';
 import { log } from './log';
 import LineByLine from 'n-readlines';
 import type { RParseRequestFromFile } from '../r-bridge/retriever';
-import type { FlowrFileProvider } from '../project/context/flowr-file';
-import type { NamespaceFormat } from '../project/plugins/file-plugins/flowr-namespace-file';
 
 /**
  * Represents a table, identified by a header and a list of rows.
@@ -175,77 +173,6 @@ export function readLineByLineSync(filePath: string, onLine: (line: Buffer, line
 export function getParentDirectory(directory: string): string{
 	// apparently this is somehow the best way to do it in node, what
 	return directory.split(path.sep).slice(0, -1).join(path.sep);
-}
-
-/**
- * Parses the given NAMESPACE file
- * @param file - The file to parse
- * @returns NamespaceFormat
- */
-export function parseNamespace(file: FlowrFileProvider): NamespaceFormat {
-	const result = {
-		current: {
-			exportedSymbols:      [] as string[],
-			exportedFunctions:    [] as string[],
-			exportS3Generics:     new Map<string, string[]>(),
-			loadsWithSideEffects: false,
-		},
-	} as NamespaceFormat;
-	const fileContent = file.content().toString().replaceAll(cleanLineCommentRegex, '').trim()
-		.split(/\r?\n/).filter(Boolean);
-
-	for(const line of fileContent) {
-		const match = line.trim().match(/^(\w+)\(([^)]*)\)$/);
-		if(!match) {
-			continue;
-		}
-		const [, type, args] = match;
-
-		switch(type) {
-			case 'exportClasses':
-			case 'exportMethods':
-				result.current.exportedFunctions.push(args);
-				break;
-			case 'S3method':
-			{
-				const parts = args.split(',').map(s => s.trim());
-				if(parts.length !== 2) {
-					continue;
-				}
-				const [pkg, func] = parts;
-				let arr = result.current.exportS3Generics.get(pkg);
-				if(!arr) {
-					arr = [];
-					result.current.exportS3Generics.set(pkg, arr);
-				}
-				arr.push(func);
-				break;
-			}
-			case 'export':
-				result.current.exportedSymbols.push(args);
-				break;
-			case 'useDynLib':
-			{
-				const parts = args.split(',').map(s => s.trim());
-				if(parts.length !== 2) {
-					continue;
-				}
-				const [pkg] = parts;
-				if(!result[pkg]) {
-					result[pkg] = {
-						exportedSymbols:      [],
-						exportedFunctions:    [],
-						exportS3Generics:     new Map<string, string[]>(),
-						loadsWithSideEffects: false,
-					};
-				}
-				result[pkg].loadsWithSideEffects = true;
-				break;
-			}
-		}
-	}
-
-	return result;
 }
 
 /**
